done tasks 1-3

+++ tasks/task6/main.c
@@ -3,30 +3,30 @@
 #include <string.h>
 
 struct Item {
-    int x;
-    char name[32];
+	int x;
+	char name[32];
 };
 
 int main(int argc, char **argv) {
-    int count = (argc > 1) ? atoi(argv[1]) : 1000000000;
-    size_t bytes = (size_t)(count * (int)sizeof(struct Item));
+	int count = (argc > 1) ? atoi(argv[1]) : 1000000000;
+	size_t bytes = (size_t)(count * sizeof(struct Item));
 
-    struct Item *ptr = malloc(4 * sizeof(struct Item));
-    if (!ptr) return 1;
+	struct Item *ptr = malloc(4 * sizeof(struct Item));
+	if (!ptr)
+		return 1;
 
-    ptr = realloc(ptr, bytes);
-    if (!ptr) {
-        printf("realloc failed\n");
-        return 1;
-    }
+	ptr = realloc(ptr, bytes);
+	if (!ptr) {
+		printf("realloc failed\n");
+		return 1;
+	}
 
-    for (int i = 0; i < count; i++) {
-        ptr[i].x = i;
-        strcpy(ptr[i].name, "example");
-    }
+	for (int i = 0; i < count; i++) {
+		ptr[i].x = i;
+		strcpy(ptr[i].name, "example");
+	}
 
-    printf("done: %d items, first=%d\n", count, ptr[0].x);
-    free(ptr);
-    return 0;
+	printf("done: %d items, first=%d\n", count, ptr[0].x);
+	free(ptr);
+	return 0;
 }
-
